Make setup.py extract version from source
Rather than hard-code the version number in two places, let setup.py
extract the __version__ string from the package's __init__.py.

Now the only places it needs to be updated when preparing a release
is __init_.py and HISTORY.rst. The latter is unfortunate but, as it
is persistent information (part of the release notes), unavoidable.