Inrease python 3 compatibility
After switching the virtual environment to use version 3 rather
than version 2, flake8 found new style errors which are now fixed.

Python 3 has a built-in exception with the same name as one declared
in gerritsite.py (ConnectionError). Therefore, the exception is
renamed to SSHConnectionError.

In Python 3, subprocess.check_output() returns a byte-like string
rather than a string which breaks many operations performed on the
output. The response is now decoded as 'utf-8' to convet it to a
string.

Also added some debug logging around site connection.