Respect virtual environment in shebang
The script would run fine when launched as 'python gsshcli.py' but
failed if simply run as a script (i.e. './gsshcli.py') due to
missing imports.

The shebang was the culprit, of course. Using 'env python3' allows
it to respect the virtual environment, and lets pip reference the
correct interpreter during installation.