Update unit tests to work with paramiko
Significant rewrites as some of the internals which I was mocking
and/or monkeypatching have changed markedly.

The biggest impact was how I simulated specific return values from
the SSH connection. Specifically, the connected_site fixture needed
some heavy lifting, which thankfully did not propogate into all the
tests that use it.