Allow init() arguments to override .ssh/config
The GerritSSHClient from pygerrit accepts values for username and
port in its creation, but the internal _configure method basically
ignored them in favor of the values found in .ssh/config.

Now, if either username or port are provided to the constructor,
they will override the value found in the configuration file.

One use might be if someone has multiple user accounts set up on
a test instance of gerrit - all with the same SSH key. Thus the
port and keyfile information can be taken from the config file,
but a different username provided to the Site constructor.