Switch to using HEAD of paramiko
The maintainers of paramiko have now merged Scott Maxwell's pull
request to support Python 3 into the master branch. It should
appear in v1.14 of paramiko.

Update the requirements file to pull from paramiko@master for now
so that we get all the fixes that were included in the merge.

Later, the requirements will need to be modified to specify >=1.14