Rename misc modules to lsprojects
With the addition of support for versioned options, the ProjectList
command has become complex enough to warrant its own module. As it
was the only miscellaneous command thus far implemented, and misc
is really not a good name for a module anyway, simply rename misc
to lsprojects.

Unit tests have also been updated to reflect the module renaming.