Final implementation of package metadata
... Hopefully

Moved the package metadata into the package as a JSON file which is
read wherever the data is needed (setup.py, docs/conf.py, and the
package itself).

Tried using execfile, but that was not supported easily across python
versions. JSON returns unicode, but that seems not to cause a problem
anywhere in 2.x.

Also modified gsshcli to use the default python rather than specifying
python3. It still respects virtual environments.

Fixes #1