Add basic online tests
Issue #3 was only discovered by running a live test against an
online Gerrit instance in my non-default Python version. This raised
issue 4 (Unit tests are not sufficient).

The new tests in test_online.py will now perform some basic online
sanity checks if the environment variable GSSH_TEST_INSTANCE is set
with the netloc for an active Gerrit instance.

Tests are not exhaustive and take time to complete, so they are
normally skipped. For now, they connect to the site; check the
returned version string; grab a list of all projects; and query
for the first three open code reviews. The list can be extended
as needed in the future, but for now that covers many of the areas
which have produced problems.

Online tests can be run as follows:

    GSSH_TEST_INSTANCE='example.gerrit.com' py.test

or

    GSSH_TEST_INSTANCE='example.gerrit.com' tox